backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
firebase-admin
google-genai
google-cloud-storage
google-cloud-translate
python-jose[cryptography]
python-multipart
